+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADConditionalAccessPolicy/MSFT_AADConditionalAccessPolicy.psm1
@@ -175,6 +175,10 @@ function Get-TargetResource
         $CloudAppSecurityType,
 
         [Parameter()]
+        [System.String]
+        $ContinuousAccessEvaluationMode,
+
+        [Parameter()]
         [System.Boolean]
         $SecureSignInSessionIsEnabled,
 
@@ -563,13 +567,15 @@ function Get-TargetResource
                 }
             }
         }
+        $CloudAppSecurityType = $null
         if ($Policy.SessionControls.CloudAppSecurity.IsEnabled)
         {
             $CloudAppSecurityType = [System.String]$Policy.SessionControls.CloudAppSecurity.CloudAppSecurityType
         }
-        else
+        $ContinuousAccessEvaluationMode = $null
+        if ($Policy.SessionControls.ContinuousAccessEvaluation.Mode)
         {
-            $CloudAppSecurityType = $null
+            $ContinuousAccessEvaluationMode = [System.String]$Policy.SessionControls.ContinuousAccessEvaluation.Mode
         }
         if ($Policy.SessionControls.SignInFrequency.IsEnabled)
         {
@@ -581,13 +587,10 @@ function Get-TargetResource
             $SignInFrequencyType = $null
             $SignInFrequencyIntervalValue = $null
         }
+        $PersistentBrowserMode = $null
         if ($Policy.SessionControls.PersistentBrowser.IsEnabled)
         {
             $PersistentBrowserMode = [System.String]$Policy.SessionControls.PersistentBrowser.Mode
-        }
-        else
-        {
-            $PersistentBrowserMode = $null
         }
         if ($Policy.Conditions.Users.IncludeGuestsOrExternalUsers.GuestOrExternalUserTypes)
         {
@@ -759,6 +762,7 @@ function Get-TargetResource
             CloudAppSecurityIsEnabled                = $false -or $Policy.SessionControls.CloudAppSecurity.IsEnabled
             #make false if undefined, true if true
             CloudAppSecurityType                     = [System.String]$Policy.SessionControls.CloudAppSecurity.CloudAppSecurityType
+            ContinuousAccessEvaluationMode           = $ContinuousAccessEvaluationMode
             SecureSignInSessionIsEnabled             = $false -or $Policy.SessionControls.SecureSignInSession.IsEnabled
             #no translation needed, return empty string array if undefined
             SignInFrequencyIsEnabled                 = $false -or $Policy.SessionControls.SignInFrequency.IsEnabled
@@ -979,6 +983,10 @@ function Set-TargetResource
         [Parameter()]
         [System.String]
         $CloudAppSecurityType,
+
+        [Parameter()]
+        [System.String]
+        $ContinuousAccessEvaluationMode,
 
         [Parameter()]
         [System.Boolean]
@@ -1819,6 +1827,7 @@ function Set-TargetResource
             $sessionControls = @{
                 applicationEnforcedRestrictions = $null
                 cloudAppSecurity                = $null
+                continuousAccessEvaluation      = $null
                 secureSignInSession             = $null
                 signInFrequency                 = $null
                 persistentBrowser               = $null
@@ -1838,6 +1847,12 @@ function Set-TargetResource
                     cloudAppSecurityType = $CloudAppSecurityType
                 }
                 $sessionControls.cloudAppSecurity = $cloudAppSecurityValue
+            }
+            if ($ContinuousAccessEvaluationMode)
+            {
+                $sessionControls.continuousAccessEvaluation = @{
+                    mode = $ContinuousAccessEvaluationMode
+                }
             }
             if ($SecureSignInSessionIsEnabled)
             {
@@ -2147,6 +2162,10 @@ function Test-TargetResource
         [Parameter()]
         [System.String]
         $CloudAppSecurityType,
+
+        [Parameter()]
+        [System.String]
+        $ContinuousAccessEvaluationMode,
 
         [Parameter()]
         [System.Boolean]